Move and split linter tests
